requirements.txt
pdfminer.six==20221105
spacy==3.5.0
email-validator==1.3.0
